feat: implement GraphicalEditorAutocomplete component for enhanced variable selection
